requirements.txt
# ── Core (always required) ─────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
torchxrayvision>=1.3.0      # TorchXRayVision: classifier + PSPNet lung segmenter
                             # Weights auto-download on first use

transformers>=4.45.0         # Grounding DINO + CheXagent-2 + Qwen2.5-VL
accelerate>=0.30.0
Pillow>=10.0.0
numpy>=1.26.0
scikit-image>=0.23.0
scipy>=1.12.0

# ── MedSAM (required for real segmentation) ───────────────────────────────────
# Install manually:
#   git clone https://github.com/bowang-lab/MedSAM && cd MedSAM && pip install -e .
# Then download checkpoint:
#   python scripts/download_models.py
# segment-anything is installed as part of the MedSAM clone above

# ── Qwen2.5-VL summarizer ─────────────────────────────────────────────────────
bitsandbytes>=0.43.0         # 4-bit quantisation (~5 GB VRAM vs ~16 GB)
qwen-vl-utils>=0.0.8         # Qwen vision preprocessing helpers

# ── Optional utilities ────────────────────────────────────────────────────────
loguru>=0.7.0
tqdm>=4.66.0
